+++ stage3/xfce/stage3-xfce.sh
@@ -569,7 +569,7 @@ systemctl enable lightdm
 cd ..
 rm -rf lightdm-gtk-greeter-2.0.9
 # Firefox.
-tar --no-same-owner -xf firefox-134.0.tar.bz2 -C /usr/lib
+tar --no-same-owner -xf firefox-134.0.2.tar.bz2 -C /usr/lib
 mkdir -p /usr/lib/firefox/distribution
 cat > /usr/lib/firefox/distribution/policies.json << END
 {